Refs #34110 -- Reorganized django.core.files.storage into a separate module.
